@@ -33,12 +33,21 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
     /** @var \Ivory\HttpAdapterBundle\DataCollector\IvoryHttpAdapterDataCollector */
     private $dataCollector;
 
+    /** @var \PHPUnit_Framework_MockObject_MockObject|\Ivory\HttpAdapter\Event\Formatter\FormatterInterface */
+    private $formatter;
+
+    /** @var \PHPUnit_Framework_MockObject_MockObject|\Ivory\HttpAdapter\Event\Timer\TimerInterface */
+    private $timer;
+
     /**
      * {@inheritdoc}
      */
     protected function setUp()
     {
-        $this->dataCollector = new IvoryHttpAdapterDataCollector();
+        $this->formatter = $this->createFormatterMock();
+        $this->timer = $this->createTimerMock();
+
+        $this->dataCollector = new IvoryHttpAdapterDataCollector($this->formatter, $this->timer);
     }
 
     /**
@@ -69,148 +78,196 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
     public function testPostSendEvent()
     {
         $httpAdapter = $this->createHttpAdapterMock();
-        $request = $this->createRequestMock();
-        $response = $this->createResponseMock();
 
-        $this->dataCollector->onPreSend($this->createPreSendEvent($httpAdapter, $request));
-        $this->dataCollector->onPostSend($this->createPostSendEvent($httpAdapter, $request, $response));
+        $this->timer
+            ->expects($this->once())
+            ->method('start')
+            ->with($this->identicalTo($request = $this->createRequestMock()))
+            ->will($this->returnValue($startedRequest = $this->createRequestMock()));
+
+        $this->timer
+            ->expects($this->once())
+            ->method('stop')
+            ->with($this->identicalTo($startedRequest))
+            ->will($this->returnValue($stoppedRequest = $this->createRequestMock()));
+
+        $this->formatter
+            ->expects($this->once())
+            ->method('formatRequest')
+            ->with($this->identicalTo($stoppedRequest))
+            ->will($this->returnValue($formattedRequest = array('request')));
+
+        $this->formatter
+            ->expects($this->once())
+            ->method('formatResponse')
+            ->with($this->identicalTo($response = $this->createResponseMock()))
+            ->will($this->returnValue($formattedResponse = array('response')));
+
+        $this->dataCollector->onPreSend($preSendEvent = $this->createPreSendEvent($httpAdapter, $request));
+        $this->dataCollector->onPostSend($this->createPostSendEvent($httpAdapter, $startedRequest, $response));
+
+        $this->assertSame($startedRequest, $preSendEvent->getRequest());
 
         $this->assertCount(1, $this->dataCollector);
-        $this->assertCount(1, $this->dataCollector->getResponses());
+        $this->assertCount(1, $responses = $this->dataCollector->getResponses());
+        $this->assertArrayHasKey(0, $responses);
         $this->assertEmpty($this->dataCollector->getExceptions());
 
-        foreach ($this->dataCollector->getResponses() as $debug) {
-            $this->assertArrayHasKey('adapter', $debug);
-            $this->assertSame('name', $debug['adapter']);
+        $response = $responses[0];
 
-            $this->assertArrayHasKey('request', $debug);
-            $this->assertSame(
-                array(
-                    'protocol_version' => $request->getProtocolVersion(),
-                    'url'              => $request->getUrl(),
-                    'method'           => $request->getMethod(),
-                    'headers'          => array('foo' => 'bar'),
-                    'raw_datas'        => 'foo=bar',
-                    'datas'            => array('baz' => 'bat'),
-                    'files'            => array('bit' => __FILE__),
-                    'parameters'       => array('time' => 0.1),
-                ),
-                $debug['request']
-            );
+        $this->assertArrayHasKey('adapter', $response);
+        $this->assertSame('name', $response['adapter']);
 
-            $this->assertArrayHasKey('response', $debug);
-            $this->assertSame(
-                array(
-                    'protocol_version' => $response->getProtocolVersion(),
-                    'status_code'      => $response->getStatusCode(),
-                    'reason_phrase'    => $response->getReasonPhrase(),
-                    'headers'          => array('bal' => 'bol'),
-                    'body'             => 'body',
-                    'parameters'       => array('bil' => 'bob'),
-                ),
-                $debug['response']
-            );
-        }
+        $this->assertArrayHasKey('request', $response);
+        $this->assertSame($formattedRequest, $response['request']);
+
+        $this->assertArrayHasKey('response', $response);
+        $this->assertSame($formattedResponse, $response['response']);
     }
 
     public function testExceptionEvent()
     {
         $httpAdapter = $this->createHttpAdapterMock();
-        $request = $this->createRequestMock();
-        $exception = $this->createExceptionMock($request);
 
-        $this->dataCollector->onPreSend($this->createPreSendEvent($httpAdapter, $request));
+        $this->timer
+            ->expects($this->once())
+            ->method('start')
+            ->with($this->identicalTo($request = $this->createRequestMock()))
+            ->will($this->returnValue($startedRequest = $this->createRequestMock()));
+
+        $this->timer
+            ->expects($this->once())
+            ->method('stop')
+            ->with($this->identicalTo($startedRequest))
+            ->will($this->returnValue($stoppedRequest = $this->createRequestMock()));
+
+        $this->formatter
+            ->expects($this->once())
+            ->method('formatRequest')
+            ->with($this->identicalTo($stoppedRequest))
+            ->will($this->returnValue($formattedRequest = array('request')));
+
+        $exception = $this->createExceptionMock($startedRequest);
+        $exception
+            ->expects($this->once())
+            ->method('setRequest')
+            ->with($this->identicalTo($stoppedRequest));
+
+        $this->formatter
+            ->expects($this->once())
+            ->method('formatException')
+            ->with($this->identicalTo($exception))
+            ->will($this->returnValue($formattedException = array('exception')));
+
+        $this->dataCollector->onPreSend($preSendEvent = $this->createPreSendEvent($httpAdapter, $request));
         $this->dataCollector->onException($this->createExceptionEvent($httpAdapter, $exception));
 
+        $this->assertSame($startedRequest, $preSendEvent->getRequest());
+
         $this->assertCount(1, $this->dataCollector);
-        $this->assertCount(1, $this->dataCollector->getExceptions());
+        $this->assertCount(1, $exceptions = $this->dataCollector->getExceptions());
+        $this->assertArrayHasKey(0, $exceptions);
         $this->assertEmpty($this->dataCollector->getResponses());
 
-        foreach ($this->dataCollector->getExceptions() as $debug) {
-            $this->assertArrayHasKey('adapter', $debug);
-            $this->assertSame('name', $debug['adapter']);
+        $exception = $exceptions[0];
 
-            $this->assertArrayHasKey('request', $debug);
-            $this->assertSame(
-                array(
-                    'protocol_version' => $request->getProtocolVersion(),
-                    'url'              => $request->getUrl(),
-                    'method'           => $request->getMethod(),
-                    'headers'          => array('foo' => 'bar'),
-                    'raw_datas'        => 'foo=bar',
-                    'datas'            => array('baz' => 'bat'),
-                    'files'            => array('bit' => __FILE__),
-                    'parameters'       => array('time' => 0.1),
-                ),
-                $debug['request']
-            );
+        $this->assertArrayHasKey('adapter', $exception);
+        $this->assertSame('name', $exception['adapter']);
 
-            $this->assertArrayHasKey('exception', $debug);
-            $this->assertSame(
-                array(
-                    'code'    => $exception->getCode(),
-                    'message' => $exception->getMessage(),
-                    'line'    => $exception->getLine(),
-                    'file'    => $exception->getFile(),
-                ),
-                $debug['exception']
-            );
-        }
+        $this->assertArrayHasKey('request', $exception);
+        $this->assertSame($formattedRequest, $exception['request']);
+
+        $this->assertArrayHasKey('exception', $exception);
+        $this->assertSame($formattedException, $exception['exception']);
     }
 
     public function testMultiPostSendEvent()
     {
         $httpAdapter = $this->createHttpAdapterMock();
 
-        $requests = array(
+        $requests = [
             $request1 = $this->createRequestMock(),
             $request2 = $this->createRequestMock(),
-        );
+        ];
 
-        $responses = array(
-            $response1 = $this->createResponseMock($request1),
-            $this->createResponseMock($request2),
-        );
+        $this->timer
+            ->expects($this->exactly(2))
+            ->method('start')
+            ->will($this->returnValueMap([
+                [$request1, $startedRequest1 = $this->createRequestMock()],
+                [$request2, $startedRequest2 = $this->createRequestMock()],
+            ]));
+
+        $this->timer
+            ->expects($this->exactly(2))
+            ->method('stop')
+            ->will($this->returnValueMap([
+                [$startedRequest1, $stoppedRequest1 = $this->createRequestMock()],
+                [$startedRequest2, $stoppedRequest2 = $this->createRequestMock()],
+            ]));
+
+        $this->formatter
+            ->expects($this->exactly(2))
+            ->method('formatRequest')
+            ->will($this->returnValueMap([
+                [$stoppedRequest1, $formattedRequest1 = ['request1']],
+                [$stoppedRequest2, $formattedRequest2 = ['request2']],
+            ]));
+
+        $responses = [
+            $response1 = $this->createResponseMock($startedRequest1),
+            $response2 = $this->createResponseMock($startedRequest2),
+        ];
+
+        $response1
+            ->expects($this->once())
+            ->method('withParameter')
+            ->with($this->identicalTo('request'), $this->identicalTo($stoppedRequest1))
+            ->will($this->returnValue($updatedResponse1 = $this->createResponseMock($stoppedRequest1)));
+
+        $response2
+            ->expects($this->once())
+            ->method('withParameter')
+            ->with($this->identicalTo('request'), $this->identicalTo($stoppedRequest2))
+            ->will($this->returnValue($updatedResponse2 = $this->createResponseMock($stoppedRequest2)));
+
+        $this->formatter
+            ->expects($this->exactly(2))
+            ->method('formatResponse')
+            ->will($this->returnValueMap([
+                [$response1, $formattedResponse1 = ['response1']],
+                [$response2, $formattedResponse2 = ['response2']],
+            ]));
 
         $this->dataCollector->onMultiPreSend($this->createMultiPreSendEvent($httpAdapter, $requests));
         $this->dataCollector->onMultiPostSend($this->createMultiPostSendEvent($httpAdapter, $responses));
 
         $this->assertCount(count($responses), $this->dataCollector);
-        $this->assertCount(count($responses), $this->dataCollector->getResponses());
+        $this->assertCount(count($responses), $responses = $this->dataCollector->getResponses());
+        $this->assertArrayHasKey(0, $responses);
+        $this->assertArrayHasKey(1, $responses);
         $this->assertEmpty($this->dataCollector->getExceptions());
 
-        foreach ($this->dataCollector->getResponses() as $debug) {
-            $this->assertArrayHasKey('adapter', $debug);
-            $this->assertSame('name', $debug['adapter']);
+        $response1 = $responses[0];
+        $response2 = $responses[1];
 
-            $this->assertArrayHasKey('request', $debug);
-            $this->assertSame(
-                array(
-                    'protocol_version' => $request1->getProtocolVersion(),
-                    'url'              => $request1->getUrl(),
-                    'method'           => $request1->getMethod(),
-                    'headers'          => array('foo' => 'bar'),
-                    'raw_datas'        => 'foo=bar',
-                    'datas'            => array('baz' => 'bat'),
-                    'files'            => array('bit' => __FILE__),
-                    'parameters'       => array('time' => 0.1),
-                ),
-                $debug['request']
-            );
+        $this->assertArrayHasKey('adapter', $response1);
+        $this->assertSame('name', $response1['adapter']);
 
-            $this->assertArrayHasKey('response', $debug);
-            $this->assertSame(
-                array(
-                    'protocol_version' => $response1->getProtocolVersion(),
-                    'status_code'      => $response1->getStatusCode(),
-                    'reason_phrase'    => $response1->getReasonPhrase(),
-                    'headers'          => array('bal' => 'bol'),
-                    'body'             => 'body',
-                    'parameters'       => array('bil' => 'bob'),
-                ),
-                $debug['response']
-            );
-        }
+        $this->assertArrayHasKey('adapter', $response2);
+        $this->assertSame('name', $response2['adapter']);
+
+        $this->assertArrayHasKey('request', $response1);
+        $this->assertSame($formattedRequest1, $response1['request']);
+
+        $this->assertArrayHasKey('request', $response2);
+        $this->assertSame($formattedRequest2, $response2['request']);
+
+        $this->assertArrayHasKey('response', $response1);
+        $this->assertSame($formattedResponse1, $response1['response']);
+
+        $this->assertArrayHasKey('response', $response2);
+        $this->assertSame($formattedResponse2, $response2['response']);
     }
 
     public function testMultiExceptionEvent()
@@ -222,57 +279,97 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
             $request2 = $this->createRequestMock(),
         );
 
+        $this->timer
+            ->expects($this->exactly(2))
+            ->method('start')
+            ->will($this->returnValueMap([
+                [$request1, $startedRequest1 = $this->createRequestMock()],
+                [$request2, $startedRequest2 = $this->createRequestMock()],
+            ]));
+
+        $this->timer
+            ->expects($this->exactly(2))
+            ->method('stop')
+            ->will($this->returnValueMap([
+                [$startedRequest1, $stoppedRequest1 = $this->createRequestMock()],
+                [$startedRequest2, $stoppedRequest2 = $this->createRequestMock()],
+            ]));
+
+        $this->formatter
+            ->expects($this->exactly(2))
+            ->method('formatRequest')
+            ->will($this->returnValueMap(array(
+                array($stoppedRequest1, $formattedRequest1 = array('request1')),
+                array($stoppedRequest2, $formattedRequest2 = array('request2')),
+            )));
+
         $exceptions = array(
-            $exception1 = $this->createExceptionMock($request1),
-            $this->createExceptionMock($request2),
+            $exception1 = $this->createExceptionMock($startedRequest1),
+            $exception2 = $this->createExceptionMock($startedRequest2),
         );
+
+        $this->formatter
+            ->expects($this->exactly(2))
+            ->method('formatException')
+            ->will($this->returnValueMap(array(
+                array($exception1, $formattedException1 = array('exception1')),
+                array($exception2, $formattedException2 = array('exception2')),
+            )));
 
         $this->dataCollector->onMultiPreSend($this->createMultiPreSendEvent($httpAdapter, $requests));
         $this->dataCollector->onMultiException($this->createMultiExceptionEvent($httpAdapter, $exceptions));
 
-        $this->assertCount(count($exceptions), $this->dataCollector);
-        $this->assertCount(count($exceptions), $this->dataCollector->getExceptions());
+        $this->assertCount(2, $this->dataCollector);
+        $this->assertCount(2, $exceptions = $this->dataCollector->getExceptions());
+        $this->assertArrayHasKey(0, $exceptions);
+        $this->assertArrayHasKey(1, $exceptions);
         $this->assertEmpty($this->dataCollector->getResponses());
 
-        foreach ($this->dataCollector->getExceptions() as $debug) {
-            $this->assertArrayHasKey('adapter', $debug);
-            $this->assertSame('name', $debug['adapter']);
+        $exception1 = $exceptions[0];
+        $exception2 = $exceptions[1];
 
-            $this->assertArrayHasKey('request', $debug);
-            $this->assertSame(
-                array(
-                    'protocol_version' => $request1->getProtocolVersion(),
-                    'url'              => $request1->getUrl(),
-                    'method'           => $request1->getMethod(),
-                    'headers'          => array('foo' => 'bar'),
-                    'raw_datas'        => 'foo=bar',
-                    'datas'            => array('baz' => 'bat'),
-                    'files'            => array('bit' => __FILE__),
-                    'parameters'       => array('time' => 0.1),
-                ),
-                $debug['request']
-            );
+        $this->assertArrayHasKey('adapter', $exception1);
+        $this->assertSame('name', $exception1['adapter']);
 
-            $this->assertArrayHasKey('exception', $debug);
-            $this->assertSame(
-                array(
-                    'code'    => $exception1->getCode(),
-                    'message' => $exception1->getMessage(),
-                    'line'    => $exception1->getLine(),
-                    'file'    => $exception1->getFile(),
-                ),
-                $debug['exception']
-            );
-        }
+        $this->assertArrayHasKey('adapter', $exception2);
+        $this->assertSame('name', $exception2['adapter']);
+
+        $this->assertArrayHasKey('request', $exception1);
+        $this->assertSame($formattedRequest1, $exception1['request']);
+
+        $this->assertArrayHasKey('request', $exception2);
+        $this->assertSame($formattedRequest2, $exception2['request']);
+
+        $this->assertArrayHasKey('exception', $exception1);
+        $this->assertSame($formattedException1, $exception1['exception']);
+
+        $this->assertArrayHasKey('exception', $exception2);
+        $this->assertSame($formattedException2, $exception2['exception']);
     }
 
     public function testSerialize()
     {
-        $this->dataCollector->onPreSend($this->createPreSendEvent());
-        $this->dataCollector->onPostSend($this->createPostSendEvent());
+        $httpAdapter = $this->createHttpAdapterMock();
 
-        $this->dataCollector->onPreSend($this->createPreSendEvent());
-        $this->dataCollector->onPostSend($this->createPostSendEvent());
+        $this->timer
+            ->expects($this->once())
+            ->method('start')
+            ->with($this->identicalTo($request = $this->createRequestMock()))
+            ->will($this->returnValue($startedRequest = $this->createRequestMock()));
+
+        $this->timer
+            ->expects($this->once())
+            ->method('stop')
+            ->with($this->identicalTo($startedRequest))
+            ->will($this->returnValue($stoppedRequest = $this->createRequestMock()));
+
+        $this->dataCollector->onPreSend($preSendEvent = $this->createPreSendEvent($httpAdapter, $request));
+
+        $this->dataCollector->onPostSend($this->createPostSendEvent(
+            $httpAdapter,
+            $preSendEvent->getRequest(),
+            $this->createResponseMock($preSendEvent->getRequest())
+        ));
 
         $dataCollector = unserialize(serialize($this->dataCollector));
 
@@ -280,6 +377,18 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
         $this->assertSame($this->dataCollector->getResponses(), $dataCollector->getResponses());
         $this->assertSame($this->dataCollector->getExceptions(), $dataCollector->getExceptions());
         $this->assertSame($this->dataCollector->getTime(), $dataCollector->getTime());
+    }
+
+    public function testSubscribedEvents()
+    {
+        $this->assertSame(array(
+            'ivory.http_adapter.pre_send'        => array('onPreSend', 100),
+            'ivory.http_adapter.post_send'       => array('onPostSend', 100),
+            'ivory.http_adapter.exception'       => array('onException', 100),
+            'ivory.http_adapter.multi_pre_send'  => array('onMultiPreSend', 100),
+            'ivory.http_adapter.multi_post_send' => array('onMultiPostSend', 100),
+            'ivory.http_adapter.multi_exception' => array('onMultiException', 100),
+        ), $this->dataCollector->getSubscribedEvents());
     }
 
     /**
@@ -347,7 +456,7 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
      *
      * @return \Ivory\HttpAdapter\Event\MultiPreSendEvent The multi pre send event.
      */
-    private function createMultiPreSendEvent(HttpAdapterInterface $httpAdapter = null, array $requests = array())
+    private function createMultiPreSendEvent(HttpAdapterInterface $httpAdapter = null, array $requests = [])
     {
         return new MultiPreSendEvent($httpAdapter ?: $this->createHttpAdapterMock(), $requests);
     }
@@ -360,7 +469,7 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
      *
      * @return \Ivory\HttpAdapter\Event\MultiPostSendEvent The multi post send event.
      */
-    private function createMultiPostSendEvent(HttpAdapterInterface $httpAdapter = null, array $responses = array())
+    private function createMultiPostSendEvent(HttpAdapterInterface $httpAdapter = null, array $responses = [])
     {
         return new MultiPostSendEvent($httpAdapter ?: $this->createHttpAdapterMock(), $responses);
     }
@@ -373,9 +482,25 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
      *
      * @return \Ivory\HttpAdapter\Event\MultiExceptionEvent The multi exception event.
      */
-    private function createMultiExceptionEvent(HttpAdapterInterface $httpAdapter = null, array $exceptions = array())
+    private function createMultiExceptionEvent(HttpAdapterInterface $httpAdapter = null, array $exceptions = [])
     {
         return new MultiExceptionEvent($httpAdapter ?: $this->createHttpAdapterMock(), $exceptions);
+    }
+
+    /**
+     * @return \PHPUnit_Framework_MockObject_MockObject|\Ivory\HttpAdapter\Event\Formatter\FormatterInterface
+     */
+    private function createFormatterMock()
+    {
+        return $this->getMock('Ivory\HttpAdapter\Event\Formatter\FormatterInterface');
+    }
+
+    /**
+     * @return \PHPUnit_Framework_MockObject_MockObject|\Ivory\HttpAdapter\Event\Timer\TimerInterface
+     */
+    private function createTimerMock()
+    {
+        return $this->getMock('Ivory\HttpAdapter\Event\Timer\TimerInterface');
     }
 
     /**
@@ -435,7 +560,7 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
         $request
             ->expects($this->any())
             ->method('getHeaders')
-            ->will($this->returnValue(array('foo' => 'bar')));
+            ->will($this->returnValue(['foo' => 'bar']));
 
         $request
             ->expects($this->any())
@@ -445,17 +570,17 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
         $request
             ->expects($this->any())
             ->method('getDatas')
-            ->will($this->returnValue(array('baz' => 'bat')));
+            ->will($this->returnValue(['baz' => 'bat']));
 
         $request
             ->expects($this->any())
             ->method('getFiles')
-            ->will($this->returnValue(array('bit' => __FILE__)));
+            ->will($this->returnValue(['bit' => __FILE__]));
 
         $request
             ->expects($this->any())
             ->method('getParameters')
-            ->will($this->returnValue(array('time' => 0.1)));
+            ->will($this->returnValue(['time' => 0.1]));
 
         return $request;
     }
@@ -488,7 +613,7 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
         $response
             ->expects($this->any())
             ->method('getHeaders')
-            ->will($this->returnValue(array('bal' => 'bol')));
+            ->will($this->returnValue(['bal' => 'bol']));
 
         $response
             ->expects($this->any())
@@ -498,7 +623,7 @@ class IvoryHttpAdapterDataCollectorTest extends \PHPUnit_Framework_TestCase
         $response
             ->expects($this->any())
             ->method('getParameters')
-            ->will($this->returnValue(array('bil' => 'bob')));
+            ->will($this->returnValue(['bil' => 'bob']));
 
         if ($request !== null) {
             $response
